+++ networkmonitor.py
@@ -10,16 +10,19 @@ from networkmonitor.tui import uiMain
 @click.command()
 @click.option('--config', default='config.json', help='json configuration file to load')
 @click.option('--newconfig', default=False, help='Generates a new configuration file')
-def init(config):
+def init(config, newconfig):
     """
         NetorkMonitor is a curses tool to monitor network nodes.
         Currently supports ping(icmp), Http: Get and Post.
 
         To get started:
-
-        networkmonitor --config "demo.json" --newconfig
+        'networkmonitor --config "demo.json" --newconfig'
 
     """
+    click.echo(f"config:{config}")
+    click.echo(f"newConfig: {newconfig}")
+
+
     main = uiMain(config)
     main.Start()
