+++ networkmonitor/__init__.py
@@ -3,7 +3,7 @@
 __version__ = '0.0.1'
 from .ping import *
 from .config import *
-from .terminal import *
+#from .terminal import *
 from .http import *
 from .cursesHelper import *
 from .monitor import *
